% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Quanjer_equations.R
\name{getQuanjerPred}
\alias{getQuanjerPred}
\title{getQuanjerPred}
\usage{
getQuanjerPred(sex, ht, age, value = "FEV1")
}
\arguments{
\item{sex}{biological sex, 'F' vs 'M'}

\item{ht}{height in meter (m)}

\item{age}{age in year}

\item{value}{PFT variables name (default in FEV1) (L)}
}
\value{
the predictive PFT values
}
\description{
Quanjer Reference Equations for calculating predictive values
for PFT variables
}
