% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Quanjer_equations.R
\name{getQuanjerULN}
\alias{getQuanjerULN}
\title{getQuanjerULN}
\usage{
getQuanjerULN(sex, ht, age, value = "FEV1")
}
\arguments{
\item{sex}{biological sex, 'F' vs 'M'}

\item{ht}{height in meter (m)}

\item{age}{age in year}

\item{value}{PFT variables name (default in FEV1) (L)}
}
\value{
the computed ULN value for the imputed PFT variable
}
\description{
Quanjer Reference Equations for calculating upper limit of norm(ULN)
values for PFT variables
}
